Minor changes.And fixes
i have used the build in function location.reload() to reload the page after the cart is emptied.
I am trying to add an event listener on the innerhtml contect on the product specific page.
Nad lastly i have created an function to redirect the user after buying an item.